refactor decryption stream initialisation
