+++ src/streams.rs
@@ -1,14 +1,12 @@
 use anyhow::Result;
 use rand::prelude::StdRng;
-
+use crate::global::crypto::DecryptStreamCiphers;
 use crate::global::{crypto::EncryptStreamCiphers, parameters::Algorithm};
-
-
-use crate::global::parameters::HeaderType;
+use crate::global::parameters::{HeaderType, Header};
 use crate::global::{SALT_LEN};
 use crate::key::{argon2_hash, gen_salt};
 use crate::secret::Secret;
-use aead::stream::EncryptorLE31;
+use aead::stream::{EncryptorLE31, DecryptorLE31};
 use aead::{NewAead};
 use aes_gcm::{Aes256Gcm, Nonce};
 use anyhow::anyhow;
@@ -17,16 +15,15 @@ use deoxys::DeoxysII256;
 use rand::{Rng, SeedableRng};
 use std::result::Result::Ok;
 
-
 pub fn init_encryption_stream(raw_key: Secret<Vec<u8>>, header_type: &HeaderType) -> Result<(EncryptStreamCiphers, [u8; SALT_LEN], Vec<u8>)> {
     let salt = gen_salt();
+    let key = argon2_hash(raw_key, &salt, &header_type.header_version)?;
 
     match header_type.algorithm {
         Algorithm::Aes256Gcm => {
             let nonce_bytes = StdRng::from_entropy().gen::<[u8; 8]>();
             let nonce = Nonce::from_slice(&nonce_bytes);
     
-            let key = argon2_hash(raw_key, &salt, &header_type.header_version)?;
             let cipher = match Aes256Gcm::new_from_slice(key.expose()) {
                 Ok(cipher) => {
                     drop(key);
@@ -45,7 +42,6 @@ pub fn init_encryption_stream(raw_key: Secret<Vec<u8>>, header_type: &HeaderType
         Algorithm::XChaCha20Poly1305 => {
             let nonce_bytes = StdRng::from_entropy().gen::<[u8; 20]>();
     
-            let key = argon2_hash(raw_key, &salt, &header_type.header_version)?;
             let cipher = match XChaCha20Poly1305::new_from_slice(key.expose()) {
                 Ok(cipher) => {
                     drop(key);
@@ -64,7 +60,6 @@ pub fn init_encryption_stream(raw_key: Secret<Vec<u8>>, header_type: &HeaderType
         Algorithm::DeoxysII256 => {
             let nonce_bytes = StdRng::from_entropy().gen::<[u8; 11]>();
     
-            let key = argon2_hash(raw_key, &salt, &header_type.header_version)?;
             let cipher = match DeoxysII256::new_from_slice(key.expose()) {
                 Ok(cipher) => {
                     drop(key);
@@ -79,6 +74,52 @@ pub fn init_encryption_stream(raw_key: Secret<Vec<u8>>, header_type: &HeaderType
                 salt,
                 nonce_bytes.to_vec(),
             ))
+        }
+    }
+}
+
+pub fn init_decryption_stream(raw_key: Secret<Vec<u8>>, header: &Header) -> Result<DecryptStreamCiphers> {
+    let key = argon2_hash(raw_key, &header.salt, &header.header_type.header_version)?;
+
+    match header.header_type.algorithm {
+        Algorithm::Aes256Gcm => {
+            let cipher = match Aes256Gcm::new_from_slice(key.expose()) {
+                Ok(cipher) => {
+                    drop(key);
+                    cipher
+                }
+                Err(_) => return Err(anyhow!("Unable to create cipher with argon2id hashed key.")),
+            };
+
+            let nonce = Nonce::from_slice(header.nonce.as_slice());
+
+            let stream = DecryptorLE31::from_aead(cipher, nonce);
+
+            Ok(DecryptStreamCiphers::Aes256Gcm(Box::new(stream)))
+        }
+        Algorithm::XChaCha20Poly1305 => {
+            let cipher = match XChaCha20Poly1305::new_from_slice(key.expose()) {
+                Ok(cipher) => {
+                    drop(key);
+                    cipher
+                }
+                Err(_) => return Err(anyhow!("Unable to create cipher with argon2id hashed key.")),
+            };
+
+            let stream = DecryptorLE31::from_aead(cipher, header.nonce.as_slice().into());
+            Ok(DecryptStreamCiphers::XChaCha(Box::new(stream)))
+        }
+        Algorithm::DeoxysII256 => {
+            let cipher = match DeoxysII256::new_from_slice(key.expose()) {
+                Ok(cipher) => {
+                    drop(key);
+                    cipher
+                }
+                Err(_) => return Err(anyhow!("Unable to create cipher with argon2id hashed key.")),
+            };
+
+            let stream = DecryptorLE31::from_aead(cipher, header.nonce.as_slice().into());
+            Ok(DecryptStreamCiphers::DeoxysII(Box::new(stream)))
         }
     }
 }
